+++ docs/src/main/tut/docs/changelog.md
@@ -8,11 +8,12 @@ title: "Changelog"
 The following sections describe major changes per version 
 and can be helpful with version upgrades.
 
-## 0.24-SNAPSHOT
+## 0.24
 
 - From now on a Corbeans "Node" controller will have to be explicitly added if desired.
 - `corbeans.nodes.xxx.testPartyName` can be used to set the Node identity for integration tests
 - Both X500 and organization names of a Node Party (`Party`) can now be used as the `nodeName` path fragment
+- Updated Spring, Spring Boot dependencies (5.1.8.RELEASE, 2.1.6.RELEASE)
 
 ## 0.23